+++ src/fr/maxlego08/menu/ZInventoryManager.java
@@ -46,6 +46,8 @@ import fr.maxlego08.menu.loader.actions.RefreshLoader;
 import fr.maxlego08.menu.loader.actions.ShopkeeperLoader;
 import fr.maxlego08.menu.loader.actions.SoundLoader;
 import fr.maxlego08.menu.loader.actions.TitleLoader;
+import fr.maxlego08.menu.loader.actions.VaultDepositLoader;
+import fr.maxlego08.menu.loader.actions.VaultWithdrawLoader;
 import fr.maxlego08.menu.loader.permissible.ItemPermissibleLoader;
 import fr.maxlego08.menu.loader.permissible.JobPermissibleLoader;
 import fr.maxlego08.menu.loader.permissible.LuckPermPermissibleLoader;
@@ -299,6 +301,10 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
         buttonManager.registerAction(new TitleLoader());
         buttonManager.registerAction(new ActionBarLoader());
         buttonManager.registerAction(new RefreshLoader());
+        if (this.plugin.isEnable(Plugins.VAULT)) {
+            buttonManager.registerAction(new VaultWithdrawLoader());
+            buttonManager.registerAction(new VaultDepositLoader());
+        }
 
         // Loading ButtonLoader
         // The first step will be to load the buttons in the plugin, so each